refactor: extract breadcrumb component & internationalization
