+++ src/components/Proposal/ProposalHero.css
@@ -32,44 +32,6 @@
   padding: 1.3em 0;
 }
 
-.Breadcrumb {
-  display: flex;
-  gap: 8px;
-}
-
-.Breadcrumb__Link {
-  color: var(--white-900);
-  text-align: center;
-  font-size: 13px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 19px;
-  text-transform: uppercase;
-  position: relative;
-  padding-left: 16px;
-}
-
-.Breadcrumb__Link--active {
-  color: var(--black-600);
-}
-
-.Breadcrumb__Link:first-child {
-  padding-left: 0;
-}
-
-.Breadcrumb__Link:before {
-  content: '\203A';
-  font-size: 22px;
-  position: absolute;
-  left: 0;
-  top: -2px;
-  opacity: 0.45;
-}
-
-.Breadcrumb__Link:first-child:before {
-  display: none;
-}
-
 .ProposalHero__Title {
   color: var(--white-900);
   font-weight: var(--weight-semi-bold);
